+++ tests/taipy/gui/gui_specific/test_df_filter.py
@@ -12,8 +12,10 @@
 import numpy as np
 import pandas as pd
 
+from taipy.gui.data.decimator.lttb import LTTB
+from taipy.gui.data.decimator.minmax import MinMaxDecimator
+from taipy.gui.data.decimator.rdp import RDP
 from taipy.gui.data.utils import _df_data_filter
-from taipy.gui.utils.decimator import LTTB, RDP, MinMaxDecimator
 
 
 def test_data_filter_1(csvdata):
